Extracting user plots into a page
