+++ src/content/applications/Users/settings/index.tsx
@@ -13,63 +13,61 @@ import SecurityTab from './SecurityTab'
 import PlotsTab from './PlotsTab'
 
 const TabsWrapper = styled(Tabs)(
-    () => `
+  () => `
     .MuiTabs-scrollableX {
       overflow-x: auto !important;
     }`
 )
 
 function ManagementUserSettings() {
-    const [currentTab, setCurrentTab] = useState<string>('plots')
+  const [currentTab, setCurrentTab] = useState<string>('activity')
 
-    const tabs = [
-        { value: 'activity', label: 'Activity' },
-        { value: 'edit_profile', label: 'Edit Profile' },
-        { value: 'notifications', label: 'Notifications' },
-        { value: 'security', label: 'Passwords/Security' },
-        { value: 'plots', label: 'My Plots' }
-    ]
+  const tabs = [
+    { value: 'activity', label: 'Activity' },
+    { value: 'edit_profile', label: 'Edit Profile' },
+    { value: 'notifications', label: 'Notifications' },
+    { value: 'security', label: 'Passwords/Security' }
+  ]
 
-    const handleTabsChange = (event: ChangeEvent<{}>, value: string): void => {
-        setCurrentTab(value)
-    }
+  const handleTabsChange = (event: ChangeEvent<{}>, value: string): void => {
+    setCurrentTab(value)
+  }
 
-    return (
-        <>
-            <Helmet>
-                <title>User Settings - Applications</title>
-            </Helmet>
-            <PageTitleWrapper>
-                <PageHeader />
-            </PageTitleWrapper>
-            <Container maxWidth='lg'>
-                <Grid container direction='row' justifyContent='center' alignItems='stretch' spacing={3}>
-                    <Grid item xs={12}>
-                        <TabsWrapper
-                            onChange={handleTabsChange}
-                            value={currentTab}
-                            variant='scrollable'
-                            scrollButtons='auto'
-                            textColor='primary'
-                            indicatorColor='primary'
-                        >
-                            {tabs.map(tab => (
-                                <Tab key={tab.value} label={tab.label} value={tab.value} />
-                            ))}
-                        </TabsWrapper>
-                    </Grid>
-                    <Grid item xs={12}>
-                        {currentTab === 'activity' && <ActivityTab />}
-                        {currentTab === 'edit_profile' && <EditProfileTab />}
-                        {currentTab === 'notifications' && <NotificationsTab />}
-                        {currentTab === 'security' && <SecurityTab />}
-                        {currentTab === 'plots' && <PlotsTab />}
-                    </Grid>
-                </Grid>
-            </Container>
-            <Footer />
-        </>
-    )
+  return (
+    <>
+      <Helmet>
+        <title>User Settings - Applications</title>
+      </Helmet>
+      <PageTitleWrapper>
+        <PageHeader />
+      </PageTitleWrapper>
+      <Container maxWidth='lg'>
+        <Grid container direction='row' justifyContent='center' alignItems='stretch' spacing={3}>
+          <Grid item xs={12}>
+            <TabsWrapper
+              onChange={handleTabsChange}
+              value={currentTab}
+              variant='scrollable'
+              scrollButtons='auto'
+              textColor='primary'
+              indicatorColor='primary'
+            >
+              {tabs.map(tab => (
+                <Tab key={tab.value} label={tab.label} value={tab.value} />
+              ))}
+            </TabsWrapper>
+          </Grid>
+          <Grid item xs={12}>
+            {currentTab === 'activity' && <ActivityTab />}
+            {currentTab === 'edit_profile' && <EditProfileTab />}
+            {currentTab === 'notifications' && <NotificationsTab />}
+            {currentTab === 'security' && <SecurityTab />}
+          </Grid>
+        </Grid>
+      </Container>
+      <Footer />
+    </>
+  )
 }
 
 export default ManagementUserSettings
